% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{arizona}
\alias{arizona}
\title{Aerial views of vegetation from Arizona, USA}
\format{
A list of logical matrices which were obtained through the 
  classification of aerial images of vegetation taken in Arizona (USA).
}
\source{
Derived from the images provided in the Supplementary Material of 
  Rodriguez et al. (2017).
}
\usage{
arizona
}
\description{
Aerial views of vegetation from Arizona, USA
}
\references{
Rodriguez, F., A. G. Mayor, M. Rietkerk, and S. Bautista. 2017. A null model 
  for assessing the cover-independent role of bare soil connectivity as 
  indicator of dryland functioning and dynamics. Ecological Indicators.
}
\keyword{datasets}
