% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{display_size_info}
\alias{display_size_info}
\alias{display_size_info.sews_result_single}
\alias{display_size_info.sews_result_list}
\alias{display_size_info.list}
\title{Display size information of a sews_result object}
\usage{
display_size_info(x, ...)

\method{display_size_info}{sews_result_single}(x, ...)

\method{display_size_info}{sews_result_list}(x, ...)

\method{display_size_info}{list}(x, ...)
}
\arguments{
\item{x}{A \code{sews_result} object (as provided by **_sews functions, such
as \code{generic_sews()}}

\item{...}{Other arguments are ignored.}
}
\description{
Display size information of a sews_result object
}
\details{
This function is used to display information about the size of matrices
  used by various functions in spatialwarnings. It is mostly destined to internal
  use.
}
