% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsw_sews.R
\name{dLSW}
\alias{dLSW}
\alias{pLSW}
\title{The Lifshitz-Slyozov-Wagner distribution}
\usage{
dLSW(x, mu, log = FALSE)

pLSW(r, mu, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{log, }{log.p logical; if \code{TRUE}, probabilities p are given as log(p)}

\item{lower.tail}{logical; if TRUE (default), probabilities are
\eqn{P[X \le x]} otherwise, \eqn{P[X > x]}}

\item{vector}{of means}
}
\value{
dLSW gives the density, pLSW gives the distribution function, both as numerical 
  vectors determined by the length of x.
}
\description{
Density and distribution function for the Lifshitz-Slyozov-Wagner (LSW)
 distribution with mean mu.
}
\details{
# TODO
}
\examples{

# Plot the density 
x <- seq(0, 10, l = 128) 
plot(x, dLSW(x, mu = 3), type = "l", col = "black")
lines(x, dLSW(x, mu = 5), type = "l", col = "red")
lines(x, dLSW(x, mu = 7), type = "l", col = "blue")
legend(x = 0, y = max(dLSW(x, mu = 3)), lty = 1, col = c("black", "red", "blue"), 
       legend = paste("mu =", c(3, 5, 7)))

}
