% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsw_sews.R
\name{lsw_sews}
\alias{lsw_sews}
\alias{raw_patch_radii_skewness}
\alias{raw_lsw_aicw}
\title{LSW-based indicators}
\usage{
lsw_sews(mat, wrap = FALSE)

raw_patch_radii_skewness(mat, wrap = FALSE)

raw_lsw_aicw(mat, wrap = FALSE)
}
\arguments{
\item{mat}{A logical matrix (TRUE/FALSE values) or a list of these}

\item{wrap}{Determines whether patches are considered to wrap around the 
matrix when reaching on of its edges}
}
\value{
\code{lsw_sews} returns an object of class \code{simple_sews_single}
  (actually a list) if \code{mat} is a single matrix or an object of class 
  \code{simple_sews_list} if \code{mat} is a list. You probably want to use some  
  of the methods written for these complicated objects instead of extracting 
  values directly (they are displayed using \code{print(<object>)}).
}
\description{
LSW indicators for systems with density-dependent aggregation
}
\details{
# TODO
}
\examples{

data(dda)
data(dda.pars)

indics <- lsw_sews(dda)
plot(indics, along = dda.pars[ ,"tau"]) 

# Compute individual indicators 

# Skewness of the distribution of patch radii
radii_skewness <- compute_indicator(dda, raw_patch_radii_skewness)
plot(radii_skewness, along = dda.pars[ ,"tau"])

# Aic weight of LSW distribution relative to a lognormal distribution 
lsw_aicw <- compute_indicator(dda, raw_lsw_aicw)
plot(lsw_aicw, along = dda.pars[ ,"tau"])



}
\references{
# TODO Siteur 2023, other relevant papers ?
}
\seealso{
dLSW, dda, raw_patch_radii_skewness, raw_lsw_aicw
}
